requirements.txt
requests==2.31.0
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
beautifulsoup4==4.12.2 